refactor(community): community board/comment 폴더 구조 변경
