+++ src/main/java/com/somemore/community/repository/board/CommunityBoardRepository.java
@@ -1,4 +1,4 @@
-package com.somemore.community.repository;
+package com.somemore.community.repository.board;
 
 import com.somemore.community.domain.CommunityBoard;
 import com.somemore.community.domain.CommunityBoardView;
