+++ src/main/java/com/somemore/community/service/board/DeleteCommunityBoardService.java
@@ -1,31 +1,32 @@
-package com.somemore.community.service;
+package com.somemore.community.service.board;
 
 import com.somemore.community.domain.CommunityBoard;
-import com.somemore.community.dto.request.CommunityBoardUpdateRequestDto;
-import com.somemore.community.repository.CommunityBoardRepository;
-import com.somemore.community.usecase.UpdateCommunityBoardUseCase;
+import com.somemore.community.repository.board.CommunityBoardRepository;
+import com.somemore.community.usecase.board.DeleteCommunityBoardUseCase;
 import com.somemore.global.exception.BadRequestException;
-
-import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import static com.somemore.global.exception.ExceptionMessage.*;
+import java.util.UUID;
 
+import static com.somemore.global.exception.ExceptionMessage.NOT_EXISTS_COMMUNITY_BOARD;
+import static com.somemore.global.exception.ExceptionMessage.UNAUTHORIZED_COMMUNITY_BOARD;
 
 @RequiredArgsConstructor
 @Transactional
 @Service
-public class UpdateCommunityBoardService implements UpdateCommunityBoardUseCase {
+class DeleteCommunityBoardService implements DeleteCommunityBoardUseCase {
 
     private final CommunityBoardRepository communityBoardRepository;
 
     @Override
-    public void updateCommunityBoard(CommunityBoardUpdateRequestDto requestDto, Long communityBoardId, UUID writerId, String imgUrl) {
-        CommunityBoard communityBoard = getCommunityBoardById(communityBoardId);
+    public void deleteCommunityBoard(UUID writerId, Long id) {
+        CommunityBoard communityBoard = getCommunityBoardById(id);
+
         validateWriter(communityBoard, writerId);
-        communityBoard.updateWith(requestDto, imgUrl);
+
+        communityBoard.markAsDeleted();
 
         communityBoardRepository.save(communityBoard);
     }
